- various improvements to package structure & descriptors - initial tests working
