+++ .circleci/config.yml
@@ -1,0 +1,20 @@
+version: 2
+jobs:
+  build:
+    working_directory: ~/tmp
+    docker:
+      - image: circleci/php:7.1-node-browsers
+    steps: # a set of executable commands
+          - checkout # special step to check out source code to working directory
+          - run: sudo composer self-update
+          - restore_cache: # special step to restore the dependency cache if `composer.lock` does not change
+              keys:
+                - composer-v1-{{ checksum "composer.lock" }}
+                # fallback to using the latest cache if no exact match is found (See https://circleci.com/docs/2.0/caching/)
+                - composer-v1-
+          - run: composer install -n --prefer-dist
+          - save_cache: # special step to save the dependency cache with the `composer.lock` cache key template
+              key: composer-v1-{{ checksum "composer.lock" }}
+              paths:
+                - vendor
+          - run: composer test
