+++ README.md
@@ -1,17 +1,35 @@
 FirePHPCore
 ===========
 
-This project holds the [FirePHP Tool](http://firephp.org) in the form of a PHP server library calld **FirePHPCore** which is compatible with the latest [Firefox Web Extension for FirePHP](https://github.com/firephp/firephp-for-firefox-devtools).
+The FirePHP server library for PHP reference implementation compatible with the latest [FirePHP Devtools Extension](https://github.com/firephp/firephp-for-browser-devtools).
 
-> FirePHP is a logging system that can display PHP variables in a browser as an application is navigated. All communication is out of band to the application which means the logging data will not interfere with the normal functioning of the application.
+> FirePHP is a logging system that can display PHP variables in a browser as an application is navigated. All communication happens via HTTP headers which means the logging data will not interfere with the normal functioning of the application.
 
 Install
 -------
+
+### Published
 
     composer require firephp/firephp-core
 
   * [packagist.org: firephp/firephp-core](https://packagist.org/packages/firephp/firephp-core)
   * [Release history & known issues](https://github.com/firephp/firephp-core/wiki)
+
+### Source
+
+Requirements:
+
+  * [php](https://www.php.net/)
+  * [composer](https://getcomposer.org/)
+
+Install dependencies:
+
+    composer install
+
+Run tests:
+  
+    composer test
+
 
 API
 ---
