+++ package.json
@@ -1,7 +1,7 @@
 {
     "uid": "https://github.com/firephp/firephp-core/",
     "name": "firephp-core",
-    "version": "0.4.0",
+    "version": "0.5.0",
     "description": "Minimal library for sending PHP variables to browsers.",
     "repositories": [
         {
