+++ tests/API/newlines.php
