--- tests/phpunit.xml.dist
@@ -1,9 +1,0 @@
-<phpunit bootstrap="TestHelper.php">
-
-  <testsuites>
-    <testsuite>
-      <directory suffix=".php">./</directory>
-    </testsuite>
-  </testsuites>
-
-</phpunit>
